+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.5.0
+### Fix
+- [GEN-26] Fixed visual bugs
+### Changes
+- [GEN-32] File management review
+  - [GEN-18] Added Catch cleanup
+  - [GEN-29] Added special catch for development
+  - [GEN-8] Added Buttons to : Clear catch, Remove local Java and Delete all data
+- [GEN-33] Make settings as popping Modal
+---
 # 0.4.71
 - Assets minimized
 - [GEN-21] Fixed AuthModule dropdown bug